+++ crates/core/src/flow/build_error.rs
@@ -1,23 +1,10 @@
 use miette::Diagnostic;
 use starbase_styles::{Style, Stylize};
-use std::io;
 use std::path::PathBuf;
 use thiserror::Error;
 
 #[derive(Error, Debug, Diagnostic)]
 pub enum ProtoBuildError {
-    #[diagnostic(code(proto::install::build::command_failed))]
-    #[error("Failed to execute command {}.", .command.style(Style::Shell))]
-    FailedCommand {
-        command: String,
-        #[source]
-        error: Box<io::Error>,
-    },
-
-    #[diagnostic(code(proto::install::build::command_failed))]
-    #[error("Command {} returned a {code} exit code.", .command.style(Style::Shell))]
-    FailedCommandNonZeroExit { command: String, code: i32 },
-
     #[diagnostic(code(proto::install::build::parse_version_failed))]
     #[error("Failed to parse version from {}.", .value.style(Style::Symbol))]
     FailedVersionParse {